Refactor histogram story
