+++ ui/components/core/src/components/ui/histogram/colors.ts
